Fixed a possible NPE in the FileUtil#forceDelete method.
git-svn-id: https://svn.apache.org/repos/asf/ant/ivy/core/trunk@1052390 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/ivy/util/FileUtil.java
@@ -309,9 +309,11 @@ public final class FileUtil {
         }
         if (file.isDirectory()) {
             File[] files = file.listFiles();
-            for (int i = 0; i < files.length; i++) {
-                if (!forceDelete(files[i])) {
-                    return false;
+            if (files != null) {
+                for (int i = 0; i < files.length; i++) {
+                    if (!forceDelete(files[i])) {
+                        return false;
+                    }
                 }
             }
         }
